(#1) fixing bug not clearing FLASH memory when copying data (see desc)
to write data (properly), flash memory has to be cleared each time. this gets complicated when we're managing 4 addresses where 3 of them need to persist throughout each clear. thus, we check all addresses that are NOT about to be set (via copy) and we store them in some stack-allocated variable, clear the flash, then write them back and finally write the new data from the clipboard.